requirements.txt
tensorflow
tensorflow-addons
gitpython
ConfigArgParse
numpy
GDAL
psycopg2
Pillow
matplotlib
scikit-learn
